fix: change gas vector to u128

+++ crates/starknet-types-rpc/src/v0_7_1/starknet_api_openrpc.rs
@@ -595,9 +595,9 @@ pub struct ExecutionResources {
 #[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
 pub struct DataAvailability {
     /// the data gas consumed by this transaction's data, 0 if it uses gas for DA
-    pub l1_data_gas: u64,
+    pub l1_data_gas: u128,
     /// the gas consumed by this transaction's data, 0 if it uses data gas for DA
-    pub l1_gas: u64,
+    pub l1_gas: u128,
 }
 
 #[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
